Revert "Merge branch 'feat/crud-user-for-admins'"
This reverts commit 8440a9db666da99764f5c229a3b913a0470b386c, reversing
changes made to a385e1e8ac2e24ffb7bb1b068fa386bedefd608c.

+++ ParkingSystem.Application/Services/UserService.cs
@@ -14,9 +14,9 @@ public class UserService : IUserService
         _userRepository = userRepository;
     }
 
-    public async Task RegisterUserAsync(string username, string hashedPassword, UserRole role)
+    public async Task RegisterUserAsync(string username, string hashedPassword)
     {
-        User user = new User (username, hashedPassword, role);
+        User user = new User (username, hashedPassword, UserRole.Member);
 
         await _userRepository.AddAsync(user);
     }
@@ -35,16 +35,11 @@ public class UserService : IUserService
         return user;
     }
 
-    public async Task<List<User>> SearchUserAsync(string searchString)
-    {
-        return await _userRepository.FindAllAsync(u => u.Username.Contains(searchString));
-    }
-
-    public async Task<bool> RemoveUserAsync(string username)
+    public async Task<bool> RemoveUserAsync(string username, string hashedPassword)
     {
         var user = await FindByUsernameAsync(username);
 
-        if (user is not null)
+        if (user is not null && user.HashedPassword == hashedPassword)
         {
             await _userRepository.RemoveAsync(user);
             return true;
